requirements.txt
# Core packages
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
pyyaml==6.0

# MLOps tools
mlflow==2.6.0
dvc==3.0.0

# API and serving
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.0.0

# Monitoring
prometheus-client==0.17.1
